containers/philips_hue_driver/requirements.txt
phue==1.1
paho-mqtt==1.4.0
